+++ packages/ui/src/components/bottom-navigation-item/bottom-navigation-item.tsx
@@ -1,0 +1,16 @@
+import { cn } from "../../utils/cn";
+import type { TBottomNavigationItem } from "./bottom-navigation-item.types";
+
+export function BottomNavigationItem({
+	active,
+	disabled,
+	className,
+	children,
+	...props
+}: TBottomNavigationItem = {}): HTMLElementTagNameMap["a"] {
+	return (
+		<a {...(props as any)} className={cn(active ? "active" : "", disabled ? "disabled" : "", className)}>
+			{children}
+		</a>
+	) as unknown as HTMLElementTagNameMap["a"];
+}