+++ packages/ui/src/components/bottom-navigation/bottom-navigation.tsx
@@ -1,0 +1,15 @@
+import { cn } from "../../utils/cn";
+import type { IBottomNavigation } from "./bottom-navigation.types";
+
+export function BottomNavigation({
+	size,
+	className,
+	children,
+	...props
+}: IBottomNavigation = {}): HTMLElementTagNameMap["div"] {
+	return (
+		<div {...(props as any)} className={cn("btm-nav", size, className)}>
+			{children}
+		</div>
+	) as unknown as HTMLElementTagNameMap["div"];
+}